image-to-code/requirements.txt
autogen-agentchat>=0.4.2
autogen-ext[openai]>=0.4.2
gradio>=4.31.0
Pillow>=10.0.0
openai>=1.30.0
langfuse>=2.36.0
openlit>=1.20.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
